First testing
Creation of the first (extremely basic and doesn't give the correct answer some of the time) iteration of the stat-line predictor. Only between two players for now, but will hopefully add some more. Also found a data base (in a .csv, reference in the readme) format which I will have to learn how to get data from and parse, but should be fun. Surprisingly happy this worked hahaha.