This doesn't work
Tried to implement adding everything to arrays, and a method of getting around the fact that the first few years of stats don't contain rebounds. This doesn't work, and yeah, I know that this is probably bad practise, but I wanted to prove that I've been trying to work on this stuff. Alright Google? You can still hire me.